requirements.txt
SQLAlchemy
alembic
fastapi[all]
psycopg2-binary # PÅ LINUX
selenium
webdriver-manager
beautifulsoup4
openai
python-dotenv
requests
pandas
pytest
uvicorn
python-multipart
pydantic